Review the building of a single IP HTTP client to select which feature to include (TLS matching, host header and SNI)
