+++ single-host-httpclient/src/test/java/com/github/nhenneaux/resilienthttpclient/singlehostclient/SingleHostHttpClientBuilderTest.java
@@ -22,20 +22,20 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class SingleHostHttpClientProviderTest {
+class SingleHostHttpClientBuilderTest {
     static {
         // Force init of the client without hostname check, otherwise it is cached
-        new SingleHostHttpClientProvider().buildSingleHostnameHttpClient("test");
+        SingleHostHttpClientBuilder.build("test");
     }
 
     @Test
     void shouldBuildSingleIpHttpClientAndWorksWithPublicWebsite() {
         // Given
-        final List<String> hosts = List.of("openjdk.java.net", "travis-ci.com", "github.com", "facebook.com", "amazon.com", "google.com");
+        final List<String> hosts = List.of("openjdk.java.net", "github.com", "twitter.com", "cloudflare.com", "facebook.com", "amazon.com", "google.com", "travis-ci.com", "en.wikipedia.org");
         for (String hostname : hosts) {
             final String ip = new DnsLookupWrapper().getInetAddressesByDnsLookUp(hostname).iterator().next().getHostAddress();
 
-            final HttpClient client = new SingleHostHttpClientProvider().buildSingleHostnameHttpClient(hostname);
+            final HttpClient client = SingleHostHttpClientBuilder.build(hostname);
 
 
             HttpRequest request = HttpRequest.newBuilder()
@@ -60,7 +60,7 @@ class SingleHostHttpClientProviderTest {
         final var hostname = "24max.de";
         final String ip = new DnsLookupWrapper().getInetAddressesByDnsLookUp(hostname).iterator().next().getHostAddress();
 
-        final HttpClient client = new SingleHostHttpClientProvider().buildSingleHostnameHttpClient(hostname);
+        final HttpClient client = SingleHostHttpClientBuilder.build(hostname);
 
 
         HttpRequest request = HttpRequest.newBuilder()
@@ -87,7 +87,7 @@ class SingleHostHttpClientProviderTest {
             var hostname = "openjdk.java.net";
             final String ip = new DnsLookupWrapper().getInetAddressesByDnsLookUp(hostname).iterator().next().getHostAddress();
 
-            final HttpClient client = new SingleHostHttpClientProvider().buildSingleHostnameHttpClient(hostname);
+            final HttpClient client = SingleHostHttpClientBuilder.build(hostname);
 
 
             HttpRequest request = HttpRequest.newBuilder()
@@ -116,7 +116,7 @@ class SingleHostHttpClientProviderTest {
         String hostname = "wrong.host.badssl.com";
         final String ip = new DnsLookupWrapper().getInetAddressesByDnsLookUp(hostname).iterator().next().getHostAddress();
 
-        final HttpClient client = new SingleHostHttpClientProvider().buildSingleHostnameHttpClient(hostname);
+        final HttpClient client = SingleHostHttpClientBuilder.build(hostname);
 
 
         HttpRequest request = HttpRequest.newBuilder()
@@ -138,7 +138,7 @@ class SingleHostHttpClientProviderTest {
         String hostname = "1000-sans.badssl.com";
         final String ip = new DnsLookupWrapper().getInetAddressesByDnsLookUp(hostname).iterator().next().getHostAddress();
 
-        final HttpClient client = new SingleHostHttpClientProvider().buildSingleHostnameHttpClient(hostname);
+        final HttpClient client = SingleHostHttpClientBuilder.builder(hostname).withTlsNameMatching().buildWithHostHeader();
 
 
         HttpRequest request = HttpRequest.newBuilder()
@@ -160,7 +160,7 @@ class SingleHostHttpClientProviderTest {
         String hostname = "no-subject.badssl.com";
         final String ip = new DnsLookupWrapper().getInetAddressesByDnsLookUp(hostname).iterator().next().getHostAddress();
 
-        final HttpClient client = new SingleHostHttpClientProvider().buildSingleHostnameHttpClient(hostname);
+        final HttpClient client = SingleHostHttpClientBuilder.builder(hostname).withTlsNameMatching().buildWithHostHeader();
 
 
         HttpRequest request = HttpRequest.newBuilder()
@@ -182,7 +182,7 @@ class SingleHostHttpClientProviderTest {
         String hostname = "no-common-name.badssl.com";
         final String ip = new DnsLookupWrapper().getInetAddressesByDnsLookUp(hostname).iterator().next().getHostAddress();
 
-        final HttpClient client = new SingleHostHttpClientProvider().buildSingleHostnameHttpClient(hostname);
+        final HttpClient client = SingleHostHttpClientBuilder.builder(hostname).withTlsNameMatching().buildWithHostHeader();
 
 
         HttpRequest request = HttpRequest.newBuilder()
@@ -213,7 +213,7 @@ class SingleHostHttpClientProviderTest {
 
     @Test
     void unreachableAddress() {
-        final HttpClient client = new SingleHostHttpClientProvider().buildSingleHostnameHttpClient("no.http.server", null, HttpClient.newBuilder().connectTimeout(Duration.ofMillis(200)));
+        final HttpClient client = SingleHostHttpClientBuilder.build("no.http.server", null, HttpClient.newBuilder().connectTimeout(Duration.ofMillis(200)));
 
         HttpRequest request = HttpRequest.newBuilder()
                 .uri(URI.create("https://10.2.3.4"))
@@ -232,7 +232,7 @@ class SingleHostHttpClientProviderTest {
 
     @Test
     void noSubjectAlternativeName() {
-        final HttpClient client = new SingleHostHttpClientProvider().buildSingleHostnameHttpClient("no.http.server", null, HttpClient.newBuilder().connectTimeout(Duration.ofMillis(1_000)));
+        final HttpClient client = SingleHostHttpClientBuilder.build("no.http.server", null, HttpClient.newBuilder().connectTimeout(Duration.ofMillis(1_000)));
 
         HttpRequest request = HttpRequest.newBuilder()
                 .uri(URI.create("https://1.1.1.1"))
@@ -255,7 +255,7 @@ class SingleHostHttpClientProviderTest {
         String hostname = "http.badssl.com";
         final String ip = new DnsLookupWrapper().getInetAddressesByDnsLookUp(hostname).iterator().next().getHostAddress();
 
-        final HttpClient client = new SingleHostHttpClientProvider().buildSingleHostnameHttpClient(hostname);
+        final HttpClient client = SingleHostHttpClientBuilder.builder(hostname).withTlsNameMatching().buildWithHostHeader();
 
 
         HttpRequest request = HttpRequest.newBuilder()
@@ -276,7 +276,7 @@ class SingleHostHttpClientProviderTest {
     @Test
     void shouldHandleNoSuchAlgorithm() {
         final NoSuchAlgorithmException noSuchAlgorithmException = new NoSuchAlgorithmException();
-        final IllegalStateException illegalStateException = assertThrows(IllegalStateException.class, () -> SingleHostHttpClientProvider.RethrowGeneralSecurityException.handleGeneralSecurityException(() -> {
+        final IllegalStateException illegalStateException = assertThrows(IllegalStateException.class, () -> SingleHostHttpClientBuilder.RethrowGeneralSecurityException.handleGeneralSecurityException(() -> {
             throw noSuchAlgorithmException;
         }));
